Deleted warnings

+++ src/include/tbs/shared/Map.h
@@ -31,7 +31,7 @@ public:
     int getSize() const;
     int getWidth() const;
     int getHeight() const;
-    double getSeed() const;
+    int getSeed() const;
     MapHeader getHeader() const;
     //void setHeader(const MapHeader& mapHeader);
